Migrating remaining calls to orchestrator storage
