+++ service/orchestrator/assessment_results_test.go
@@ -27,21 +27,32 @@ package orchestrator
 
 import (
 	"context"
+	"errors"
+	"fmt"
+	"io"
+	"reflect"
+	"runtime"
+	"sync"
 	"testing"
 	"time"
 
 	"clouditor.io/clouditor/api/assessment"
+	"clouditor.io/clouditor/api/orchestrator"
+	"clouditor.io/clouditor/internal/testdata"
 	"clouditor.io/clouditor/internal/testutil"
 	"clouditor.io/clouditor/internal/util"
 	"clouditor.io/clouditor/persistence"
 	"clouditor.io/clouditor/service"
+	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/metadata"
+	"google.golang.org/protobuf/types/known/emptypb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 func TestService_ListAssessmentResults(t *testing.T) {
 	type fields struct {
-		results map[string]*assessment.AssessmentResult
 		storage persistence.Storage
 		authz   service.AuthorizationStrategy
 	}
@@ -59,7 +70,7 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "request is missing",
 			fields: fields{
-				results: map[string]*assessment.AssessmentResult{},
+				storage: testutil.NewInMemoryStorage(t),
 				authz:   &service.AuthorizationStrategyAllowAll{},
 			},
 			args:    args{},
@@ -71,58 +82,62 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "request is empty",
 			fields: fields{
-				results: map[string]*assessment.AssessmentResult{},
+				storage: testutil.NewInMemoryStorage(t),
 				authz:   &service.AuthorizationStrategyAllowAll{},
 			},
 			args: args{
 				req: &assessment.ListAssessmentResultsRequest{},
 			},
-			wantRes: &assessment.ListAssessmentResultsResponse{},
+			wantRes: &assessment.ListAssessmentResultsResponse{
+				Results: []*assessment.AssessmentResult{},
+			},
 			wantErr: assert.NoError,
 		},
 		{
 			name: "list all with allow all",
 			fields: fields{
-				results: map[string]*assessment.AssessmentResult{
-					"1": {Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0))},
-					"2": {Id: "2", Timestamp: timestamppb.New(time.Unix(0, 0))},
-				},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(&assessment.AssessmentResult{Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0))}))
+					assert.NoError(t, s.Create(&assessment.AssessmentResult{Id: "2", Timestamp: timestamppb.New(time.Unix(1, 0))}))
+				}),
 				authz: &service.AuthorizationStrategyAllowAll{},
 			},
 			args: args{req: &assessment.ListAssessmentResultsRequest{}},
 			wantRes: &assessment.ListAssessmentResultsResponse{
 				Results: []*assessment.AssessmentResult{
 					{Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0))},
-					{Id: "2", Timestamp: timestamppb.New(time.Unix(0, 0))},
+					{Id: "2", Timestamp: timestamppb.New(time.Unix(1, 0))},
 				},
 			},
 			wantErr: assert.NoError,
 		},
 		{
-			name: "list all denied",
+			name: "list all only with allowed cloud service",
 			fields: fields{
-				results: map[string]*assessment.AssessmentResult{
-					"1": {Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0))},
-					"2": {Id: "2", Timestamp: timestamppb.New(time.Unix(0, 0))},
-				},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(&assessment.AssessmentResult{Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testdata.MockCloudServiceID}))
+					assert.NoError(t, s.Create(&assessment.AssessmentResult{Id: "2", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testdata.MockAnotherCloudServiceID}))
+				}),
 				authz: &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
 				req: &assessment.ListAssessmentResultsRequest{},
 			},
-			wantRes: nil,
-			wantErr: func(tt assert.TestingT, err error, i ...interface{}) bool {
-				return assert.Equal(t, err, service.ErrPermissionDenied)
+			wantRes: &assessment.ListAssessmentResultsResponse{
+				Results: []*assessment.AssessmentResult{
+					{Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testdata.MockCloudServiceID},
+				},
 			},
+			wantErr: assert.NoError,
 		},
 		{
 			name: "specify filtered cloud service ID which is not allowed",
 			fields: fields{
-				results: map[string]*assessment.AssessmentResult{
-					"1": {Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testutil.TestCloudService1},
-					"2": {Id: "2", Timestamp: timestamppb.New(time.Unix(0, 0)), CloudServiceId: testutil.TestCloudService2},
-				},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(&assessment.AssessmentResult{Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testdata.MockCloudServiceID}))
+					assert.NoError(t, s.Create(&assessment.AssessmentResult{Id: "2", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testdata.MockAnotherCloudServiceID}))
+				}),
 				authz: &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
 			},
 			args: args{
@@ -139,10 +154,10 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "return filtered cloud service ID",
 			fields: fields{
-				results: map[string]*assessment.AssessmentResult{
-					"1": {Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testutil.TestCloudService1},
-					"2": {Id: "2", Timestamp: timestamppb.New(time.Unix(0, 0)), CloudServiceId: testutil.TestCloudService2},
-				},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(&assessment.AssessmentResult{Id: "1", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testdata.MockCloudServiceID}))
+					assert.NoError(t, s.Create(&assessment.AssessmentResult{Id: "2", Timestamp: timestamppb.New(time.Unix(1, 0)), CloudServiceId: testdata.MockAnotherCloudServiceID}))
+				}),
 				authz: &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
 			},
 			args: args{
@@ -161,8 +176,10 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "return filtered cloud service ID and filtered compliant assessment results",
 			fields: fields{
-				results: getAssessmentResults(),
-				authz:   &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(getAssessmentResults()))
+				}),
+				authz: &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
@@ -187,8 +204,10 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "return filtered cloud service ID and filtered non-compliant assessment results",
 			fields: fields{
-				results: getAssessmentResults(),
-				authz:   &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(getAssessmentResults()))
+				}),
+				authz: &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
@@ -213,8 +232,11 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "return filtered compliant assessment results",
 			fields: fields{
-				results: getAssessmentResults(),
-				authz:   &service.AuthorizationStrategyAllowAll{},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(getAssessmentResults()))
+
+				}),
+				authz: &service.AuthorizationStrategyAllowAll{},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
@@ -245,8 +267,11 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "return filtered non-compliant assessment results",
 			fields: fields{
-				results: getAssessmentResults(),
-				authz:   &service.AuthorizationStrategyAllowAll{},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(getAssessmentResults()))
+
+				}),
+				authz: &service.AuthorizationStrategyAllowAll{},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
@@ -277,8 +302,11 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "return filtered cloud service ID and one filtered metric ID",
 			fields: fields{
-				results: getAssessmentResults(),
-				authz:   &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(getAssessmentResults()))
+
+				}),
+				authz: &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
@@ -303,8 +331,11 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "return filtered cloud service ID and two filtered metric IDs",
 			fields: fields{
-				results: getAssessmentResults(),
-				authz:   &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(getAssessmentResults()))
+
+				}),
+				authz: &service.AuthorizationStrategyJWT{Key: testutil.TestCustomClaims},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
@@ -336,8 +367,11 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "return one filtered metric ID",
 			fields: fields{
-				results: getAssessmentResults(),
-				authz:   &service.AuthorizationStrategyAllowAll{},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(getAssessmentResults()))
+
+				}),
+				authz: &service.AuthorizationStrategyAllowAll{},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
@@ -368,8 +402,11 @@ func TestService_ListAssessmentResults(t *testing.T) {
 		{
 			name: "Invalid cloud service id request",
 			fields: fields{
-				results: getAssessmentResults(),
-				authz:   &service.AuthorizationStrategyAllowAll{},
+				storage: testutil.NewInMemoryStorage(t, func(s persistence.Storage) {
+					assert.NoError(t, s.Create(getAssessmentResults()))
+
+				}),
+				authz: &service.AuthorizationStrategyAllowAll{},
 			},
 			args: args{
 				ctx: testutil.TestContextOnlyService1,
@@ -387,7 +424,6 @@ func TestService_ListAssessmentResults(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			svc := &Service{
-				results: tt.fields.results,
 				storage: tt.fields.storage,
 				authz:   tt.fields.authz,
 			}
@@ -397,41 +433,36 @@ func TestService_ListAssessmentResults(t *testing.T) {
 			if tt.wantRes == nil {
 				assert.Nil(t, gotRes)
 			} else {
-				for _, elem := range gotRes.Results {
-					assert.Contains(t, tt.wantRes.Results, elem)
-				}
-
-				assert.Equal(t, len(gotRes.Results), len(tt.wantRes.Results))
-
+				assert.Equal(t, tt.wantRes, gotRes)
 			}
 		})
 	}
 }
 
-func getAssessmentResults() (results map[string]*assessment.AssessmentResult) {
-	results = map[string]*assessment.AssessmentResult{
-		"1": {
+func getAssessmentResults() (results []*assessment.AssessmentResult) {
+	results = []*assessment.AssessmentResult{
+		{
 			Id:             "1",
 			Timestamp:      timestamppb.New(time.Unix(1, 0)),
 			MetricId:       "TestMetricID-1",
 			Compliant:      true,
 			CloudServiceId: testutil.TestCloudService1,
 		},
-		"2": {
+		{
 			Id:             "2",
 			Timestamp:      timestamppb.New(time.Unix(1, 0)),
 			MetricId:       "TestMetricID-2",
 			Compliant:      true,
 			CloudServiceId: testutil.TestCloudService2,
 		},
-		"3": {
+		{
 			Id:             "3",
 			Timestamp:      timestamppb.New(time.Unix(1, 0)),
 			MetricId:       "TestMetricID-2",
 			Compliant:      false,
 			CloudServiceId: testutil.TestCloudService1,
 		},
-		"4": {
+		{
 			Id:             "4",
 			Timestamp:      timestamppb.New(time.Unix(1, 0)),
 			MetricId:       "TestMetricID-1",
@@ -441,4 +472,482 @@ func getAssessmentResults() (results map[string]*assessment.AssessmentResult) {
 	}
 
 	return
+}
+
+func TestAssessmentResultHook(t *testing.T) {
+	var (
+		hookCallCounter = 0
+		wg              sync.WaitGroup
+	)
+	wg.Add(2)
+
+	firstHookFunction := func(assessmentResult *assessment.AssessmentResult, err error) {
+		hookCallCounter++
+		log.Println("Hello from inside the firstHookFunction")
+
+		wg.Done()
+	}
+
+	secondHookFunction := func(assessmentResult *assessment.AssessmentResult, err error) {
+		hookCallCounter++
+		log.Println("Hello from inside the secondHookFunction")
+
+		wg.Done()
+	}
+
+	service := NewService()
+	service.RegisterAssessmentResultHook(firstHookFunction)
+	service.RegisterAssessmentResultHook(secondHookFunction)
+
+	// Check if first hook is registered
+	funcName1 := runtime.FuncForPC(reflect.ValueOf(service.AssessmentResultHooks[0]).Pointer()).Name()
+	funcName2 := runtime.FuncForPC(reflect.ValueOf(firstHookFunction).Pointer()).Name()
+	assert.Equal(t, funcName1, funcName2)
+
+	// Check if second hook is registered
+	funcName1 = runtime.FuncForPC(reflect.ValueOf(service.AssessmentResultHooks[1]).Pointer()).Name()
+	funcName2 = runtime.FuncForPC(reflect.ValueOf(secondHookFunction).Pointer()).Name()
+	assert.Equal(t, funcName1, funcName2)
+
+	// Check GRPC call
+	type args struct {
+		in0        context.Context
+		assessment *orchestrator.StoreAssessmentResultRequest
+	}
+	tests := []struct {
+		name     string
+		args     args
+		wantResp *orchestrator.StoreAssessmentResultResponse
+		wantErr  bool
+	}{
+		{
+			name: "Store first assessment result to the map",
+			args: args{
+				in0: context.TODO(),
+				assessment: &orchestrator.StoreAssessmentResultRequest{
+					Result: &assessment.AssessmentResult{
+						Id:             testdata.MockAssessmentResultID,
+						MetricId:       testdata.MockMetricID,
+						EvidenceId:     testdata.MockEvidenceID,
+						CloudServiceId: testdata.MockCloudServiceID,
+						Timestamp:      timestamppb.Now(),
+						MetricConfiguration: &assessment.MetricConfiguration{
+							TargetValue:    toStruct(1.0),
+							Operator:       ">=",
+							IsDefault:      true,
+							CloudServiceId: testdata.MockCloudServiceID,
+							MetricId:       testdata.MockMetricID,
+						},
+						NonComplianceComments: testdata.MockAssessmentResultNonComplianceComment,
+						Compliant:             true,
+						ResourceId:            testdata.MockResourceID,
+						ResourceTypes:         []string{"ResourceType"},
+					},
+				},
+			},
+			wantErr:  false,
+			wantResp: &orchestrator.StoreAssessmentResultResponse{},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			hookCallCounter = 0
+			s := service
+			gotResp, err := s.StoreAssessmentResult(tt.args.in0, tt.args.assessment)
+
+			// wait for all hooks (2 hooks)
+			wg.Wait()
+
+			if (err != nil) != tt.wantErr {
+				t.Errorf("StoreAssessmentResult() error = %v, wantErrMessage %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(gotResp, tt.wantResp) {
+				t.Errorf("StoreAssessmentResult() gotResp = %v, want %v", gotResp, tt.wantResp)
+			}
+
+			var results []*assessment.AssessmentResult
+			assert.NoError(t, s.storage.List(&results, "", true, 0, -1))
+
+			assert.NotEmpty(t, results)
+			assert.Equal(t, 2, hookCallCounter)
+		})
+	}
+}
+
+func TestStoreAssessmentResult(t *testing.T) {
+	type args struct {
+		in0        context.Context
+		assessment *orchestrator.StoreAssessmentResultRequest
+	}
+	tests := []struct {
+		name     string
+		args     args
+		wantResp *orchestrator.StoreAssessmentResultResponse
+		wantErr  assert.ErrorAssertionFunc
+	}{
+		{
+			name: "Store assessment to the map",
+			args: args{
+				in0: context.TODO(),
+				assessment: &orchestrator.StoreAssessmentResultRequest{
+					Result: &assessment.AssessmentResult{
+						Id:             testdata.MockAssessmentResultID,
+						MetricId:       "assessmentResultMetricID",
+						EvidenceId:     testdata.MockEvidenceID,
+						CloudServiceId: testdata.MockCloudServiceID,
+						Timestamp:      timestamppb.Now(),
+						MetricConfiguration: &assessment.MetricConfiguration{
+							TargetValue:    toStruct(1.0),
+							Operator:       "<=",
+							IsDefault:      true,
+							CloudServiceId: testdata.MockCloudServiceID,
+							MetricId:       testdata.MockMetricID,
+						},
+						NonComplianceComments: testdata.MockAssessmentResultNonComplianceComment,
+						Compliant:             true,
+						ResourceId:            testdata.MockResourceID,
+						ResourceTypes:         []string{"ResourceType"},
+					},
+				},
+			},
+			wantErr:  assert.NoError,
+			wantResp: &orchestrator.StoreAssessmentResultResponse{},
+		},
+		{
+			name: "Store assessment without metricId to the map",
+			args: args{
+				in0: context.TODO(),
+				assessment: &orchestrator.StoreAssessmentResultRequest{
+					Result: &assessment.AssessmentResult{
+						Id:             testdata.MockAssessmentResultID,
+						EvidenceId:     testdata.MockEvidenceID,
+						CloudServiceId: testdata.MockCloudServiceID,
+						Timestamp:      timestamppb.Now(),
+						MetricConfiguration: &assessment.MetricConfiguration{
+							TargetValue:    toStruct(1.0),
+							Operator:       "<=",
+							IsDefault:      true,
+							CloudServiceId: testdata.MockCloudServiceID,
+							MetricId:       testdata.MockMetricID,
+						},
+						NonComplianceComments: testdata.MockAssessmentResultNonComplianceComment,
+						Compliant:             true,
+						ResourceId:            testdata.MockResourceID,
+						ResourceTypes:         []string{"ResourceType"},
+					},
+				},
+			},
+			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
+				return assert.ErrorContains(t, err, "caused by: invalid AssessmentResult.MetricId: value length must be at least 1 runes")
+			},
+			wantResp: nil,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := NewService()
+			gotResp, err := s.StoreAssessmentResult(tt.args.in0, tt.args.assessment)
+			tt.wantErr(t, err)
+			assert.Equal(t, tt.wantResp, gotResp)
+
+			var results []*assessment.AssessmentResult
+			assert.NoError(t, s.storage.List(&results, "", true, 0, -1))
+
+			if err == nil {
+				assert.NotNil(t, results)
+			} else {
+				assert.Empty(t, results)
+			}
+		})
+	}
+}
+
+func TestStoreAssessmentResults(t *testing.T) {
+	const (
+		count1 = 1
+		count2 = 2
+	)
+
+	type fields struct {
+		countElementsInMock    int
+		countElementsInResults int
+	}
+
+	type args struct {
+		streamToServer            *mockStreamer
+		streamToClientWithSendErr *mockStreamerWithSendErr
+		streamToServerWithRecvErr *mockStreamerWithRecvErr
+	}
+
+	tests := []struct {
+		name            string
+		fields          fields
+		args            args
+		wantErr         bool
+		wantRespMessage []orchestrator.StoreAssessmentResultResponse
+		wantErrMessage  string
+	}{
+		{
+			name: "Store 2 assessment results to the map",
+			fields: fields{
+				countElementsInMock:    count2,
+				countElementsInResults: count2,
+			},
+			args:    args{streamToServer: createMockStream(createStoreAssessmentResultRequestsMock(count2))},
+			wantErr: false,
+			wantRespMessage: []orchestrator.StoreAssessmentResultResponse{
+				{
+					Status: true,
+				},
+				{
+					Status: true,
+				},
+			},
+		},
+		{
+			name: "Missing MetricID",
+			fields: fields{
+				countElementsInMock:    count1,
+				countElementsInResults: 0,
+			},
+			args:    args{streamToServer: createMockStream(createStoreAssessmentResultRequestMockWithMissingMetricID(count1))},
+			wantErr: false,
+			wantRespMessage: []orchestrator.StoreAssessmentResultResponse{
+				{
+					Status:        false,
+					StatusMessage: "MetricId: value length must be at least 1 runes",
+				},
+			},
+		},
+		{
+			name: "Error in stream to server - Recv()-err",
+			fields: fields{
+				countElementsInMock:    count1,
+				countElementsInResults: 0,
+			},
+			args:           args{streamToServerWithRecvErr: createMockStreamWithRecvErr(createStoreAssessmentResultRequestsMock(count1))},
+			wantErr:        true,
+			wantErrMessage: "rpc error: code = Unknown desc = cannot receive stream request",
+		},
+		{
+			name: "Error in stream to client - Send()-err",
+			fields: fields{
+				countElementsInMock:    count1,
+				countElementsInResults: 0,
+			},
+			args:           args{streamToClientWithSendErr: createMockStreamWithSendErr(createStoreAssessmentResultRequestsMock(count1))},
+			wantErr:        true,
+			wantErrMessage: "rpc error: code = Unknown desc = cannot stream response to the client",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := NewService()
+
+			var err error
+
+			if tt.args.streamToServer != nil {
+				err = s.StoreAssessmentResults(tt.args.streamToServer)
+			} else if tt.args.streamToClientWithSendErr != nil {
+				err = s.StoreAssessmentResults(tt.args.streamToClientWithSendErr)
+			} else if tt.args.streamToServerWithRecvErr != nil {
+				err = s.StoreAssessmentResults(tt.args.streamToServerWithRecvErr)
+			}
+
+			var results []*assessment.AssessmentResult
+			assert.NoError(t, s.storage.List(&results, "", true, 0, -1))
+
+			if (err != nil) != tt.wantErr {
+				t.Errorf("Got StoreAssessmentResults() error = %v, wantErrMessage %v", err, tt.wantErr)
+				assert.Equal(t, tt.fields.countElementsInResults, len(results))
+				return
+			} else if tt.wantErr {
+				assert.Contains(t, err.Error(), tt.wantErrMessage)
+			} else {
+				// Close stream for testing
+				close(tt.args.streamToServer.SentFromServer)
+				assert.Nil(t, err)
+				assert.Equal(t, tt.fields.countElementsInResults, len(results))
+
+				// Check all stream responses from server to client
+				i := 0
+				for elem := range tt.args.streamToServer.SentFromServer {
+					assert.Contains(t, elem.StatusMessage, tt.wantRespMessage[i].StatusMessage)
+					assert.Equal(t, tt.wantRespMessage[i].Status, elem.Status)
+					i++
+				}
+			}
+
+		})
+	}
+}
+
+// createStoreAssessmentResultRequestMockWithMissingMetricID create one StoreAssessmentResultRequest without ToolID
+func createStoreAssessmentResultRequestMockWithMissingMetricID(count int) []*orchestrator.StoreAssessmentResultRequest {
+	req := createStoreAssessmentResultRequestsMock(count)
+
+	req[0].Result.MetricId = ""
+
+	return req
+}
+
+// createStoreAssessmentResultrequestMocks creates store assessment result requests with random assessment result IDs
+func createStoreAssessmentResultRequestsMock(count int) []*orchestrator.StoreAssessmentResultRequest {
+	var mockRequests []*orchestrator.StoreAssessmentResultRequest
+
+	for i := 0; i < count; i++ {
+		storeAssessmentResultRequest := &orchestrator.StoreAssessmentResultRequest{
+			Result: &assessment.AssessmentResult{
+				Id:             uuid.NewString(),
+				MetricId:       fmt.Sprintf("assessmentResultMetricID-%d", i),
+				EvidenceId:     testdata.MockEvidenceID,
+				CloudServiceId: testdata.MockCloudServiceID,
+				Timestamp:      timestamppb.Now(),
+				MetricConfiguration: &assessment.MetricConfiguration{
+					TargetValue:    toStruct(1.0),
+					Operator:       "<=",
+					IsDefault:      true,
+					CloudServiceId: testdata.MockCloudServiceID,
+					MetricId:       testdata.MockMetricID,
+				},
+				NonComplianceComments: testdata.MockAssessmentResultNonComplianceComment,
+				Compliant:             true,
+				ResourceId:            testdata.MockResourceID,
+				ResourceTypes:         []string{"ResourceType"},
+			},
+		}
+
+		mockRequests = append(mockRequests, storeAssessmentResultRequest)
+	}
+
+	return mockRequests
+}
+
+type mockStreamer struct {
+	grpc.ServerStream
+	RecvToServer   chan *orchestrator.StoreAssessmentResultRequest
+	SentFromServer chan *orchestrator.StoreAssessmentResultResponse
+}
+
+func createMockStream(requests []*orchestrator.StoreAssessmentResultRequest) *mockStreamer {
+	m := &mockStreamer{
+		RecvToServer: make(chan *orchestrator.StoreAssessmentResultRequest, len(requests)),
+	}
+	for _, req := range requests {
+		m.RecvToServer <- req
+	}
+
+	m.SentFromServer = make(chan *orchestrator.StoreAssessmentResultResponse, len(requests))
+	return m
+}
+
+func (m mockStreamer) Send(response *orchestrator.StoreAssessmentResultResponse) error {
+	m.SentFromServer <- response
+	return nil
+}
+
+func (m mockStreamer) Recv() (*orchestrator.StoreAssessmentResultRequest, error) {
+	if len(m.RecvToServer) == 0 {
+		return nil, io.EOF
+	}
+	req, more := <-m.RecvToServer
+	if !more {
+		return nil, errors.New("empty")
+	}
+
+	return req, nil
+}
+
+func (mockStreamer) SendAndClose(_ *emptypb.Empty) error {
+	return nil
+}
+
+func (mockStreamer) SetHeader(_ metadata.MD) error {
+	panic("implement me")
+}
+
+func (mockStreamer) SendHeader(_ metadata.MD) error {
+	panic("implement me")
+}
+
+func (mockStreamer) SetTrailer(_ metadata.MD) {
+	panic("implement me")
+}
+
+func (mockStreamer) Context() context.Context {
+	panic("implement me")
+}
+
+func (mockStreamer) SendMsg(_ interface{}) error {
+	panic("implement me")
+}
+
+func (mockStreamer) RecvMsg(_ interface{}) error {
+	panic("implement me")
+}
+
+type mockStreamerWithSendErr struct {
+	grpc.ServerStream
+	RecvToServer   chan *orchestrator.StoreAssessmentResultRequest
+	SentFromServer chan *orchestrator.StoreAssessmentResultResponse
+}
+
+func (mockStreamerWithSendErr) Send(*orchestrator.StoreAssessmentResultResponse) error {
+	return errors.New("Send()-err")
+}
+
+func (m mockStreamerWithSendErr) Recv() (*orchestrator.StoreAssessmentResultRequest, error) {
+	if len(m.RecvToServer) == 0 {
+		return nil, io.EOF
+	}
+	req, more := <-m.RecvToServer
+	if !more {
+		return nil, errors.New("empty")
+	}
+
+	return req, nil
+}
+
+func createMockStreamWithSendErr(requests []*orchestrator.StoreAssessmentResultRequest) *mockStreamerWithSendErr {
+	m := &mockStreamerWithSendErr{
+		RecvToServer: make(chan *orchestrator.StoreAssessmentResultRequest, len(requests)),
+	}
+	for _, req := range requests {
+		m.RecvToServer <- req
+	}
+
+	m.SentFromServer = make(chan *orchestrator.StoreAssessmentResultResponse, len(requests))
+	return m
+}
+
+type mockStreamerWithRecvErr struct {
+	grpc.ServerStream
+	RecvToServer   chan *orchestrator.StoreAssessmentResultRequest
+	SentFromServer chan *orchestrator.StoreAssessmentResultResponse
+}
+
+func (mockStreamerWithRecvErr) Send(*orchestrator.StoreAssessmentResultResponse) error {
+	panic("implement me")
+}
+
+func (mockStreamerWithRecvErr) Recv() (*orchestrator.StoreAssessmentResultRequest, error) {
+	err := errors.New("Recv()-error")
+
+	return nil, err
+}
+
+func createMockStreamWithRecvErr(requests []*orchestrator.StoreAssessmentResultRequest) *mockStreamerWithRecvErr {
+	m := &mockStreamerWithRecvErr{
+		RecvToServer: make(chan *orchestrator.StoreAssessmentResultRequest, len(requests)),
+	}
+	for _, req := range requests {
+		m.RecvToServer <- req
+	}
+
+	m.SentFromServer = make(chan *orchestrator.StoreAssessmentResultResponse, len(requests))
+	return m
 }
