+++ service/orchestrator/orchestrator_test.go
@@ -28,9 +28,6 @@ package orchestrator
 import (
 	"context"
 	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
 	"os"
 	"reflect"
 	"runtime"
@@ -39,496 +36,22 @@ import (
 	"testing"
 
 	"clouditor.io/clouditor/api"
-	"clouditor.io/clouditor/api/assessment"
 	"clouditor.io/clouditor/api/orchestrator"
 	"clouditor.io/clouditor/internal/testdata"
 	"clouditor.io/clouditor/internal/testutil/clitest"
 	"clouditor.io/clouditor/internal/testutil/orchestratortest"
 	"clouditor.io/clouditor/persistence/inmemory"
-	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
-	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/emptypb"
 	"google.golang.org/protobuf/types/known/structpb"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 func TestMain(m *testing.M) {
 	clitest.AutoChdir()
 
 	os.Exit(m.Run())
-}
-
-func TestAssessmentResultHook(t *testing.T) {
-	var (
-		hookCallCounter = 0
-		wg              sync.WaitGroup
-	)
-	wg.Add(2)
-
-	firstHookFunction := func(assessmentResult *assessment.AssessmentResult, err error) {
-		hookCallCounter++
-		log.Println("Hello from inside the firstHookFunction")
-
-		wg.Done()
-	}
-
-	secondHookFunction := func(assessmentResult *assessment.AssessmentResult, err error) {
-		hookCallCounter++
-		log.Println("Hello from inside the secondHookFunction")
-
-		wg.Done()
-	}
-
-	service := NewService()
-	service.RegisterAssessmentResultHook(firstHookFunction)
-	service.RegisterAssessmentResultHook(secondHookFunction)
-
-	// Check if first hook is registered
-	funcName1 := runtime.FuncForPC(reflect.ValueOf(service.AssessmentResultHooks[0]).Pointer()).Name()
-	funcName2 := runtime.FuncForPC(reflect.ValueOf(firstHookFunction).Pointer()).Name()
-	assert.Equal(t, funcName1, funcName2)
-
-	// Check if second hook is registered
-	funcName1 = runtime.FuncForPC(reflect.ValueOf(service.AssessmentResultHooks[1]).Pointer()).Name()
-	funcName2 = runtime.FuncForPC(reflect.ValueOf(secondHookFunction).Pointer()).Name()
-	assert.Equal(t, funcName1, funcName2)
-
-	// Check GRPC call
-	type args struct {
-		in0        context.Context
-		assessment *orchestrator.StoreAssessmentResultRequest
-	}
-	tests := []struct {
-		name     string
-		args     args
-		wantResp *orchestrator.StoreAssessmentResultResponse
-		wantErr  bool
-	}{
-		{
-			name: "Store first assessment result to the map",
-			args: args{
-				in0: context.TODO(),
-				assessment: &orchestrator.StoreAssessmentResultRequest{
-					Result: &assessment.AssessmentResult{
-						Id:             testdata.MockAssessmentResultID,
-						MetricId:       testdata.MockMetricID,
-						EvidenceId:     testdata.MockEvidenceID,
-						CloudServiceId: testdata.MockCloudServiceID,
-						Timestamp:      timestamppb.Now(),
-						MetricConfiguration: &assessment.MetricConfiguration{
-							TargetValue:    toStruct(1.0),
-							Operator:       ">=",
-							IsDefault:      true,
-							CloudServiceId: testdata.MockCloudServiceID,
-							MetricId:       testdata.MockMetricID,
-						},
-						NonComplianceComments: testdata.MockAssessmentResultNonComplianceComment,
-						Compliant:             true,
-						ResourceId:            testdata.MockResourceID,
-						ResourceTypes:         []string{"ResourceType"},
-					},
-				},
-			},
-			wantErr:  false,
-			wantResp: &orchestrator.StoreAssessmentResultResponse{},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			hookCallCounter = 0
-			s := service
-			gotResp, err := s.StoreAssessmentResult(tt.args.in0, tt.args.assessment)
-
-			// wait for all hooks (2 hooks)
-			wg.Wait()
-
-			if (err != nil) != tt.wantErr {
-				t.Errorf("StoreAssessmentResult() error = %v, wantErrMessage %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(gotResp, tt.wantResp) {
-				t.Errorf("StoreAssessmentResult() gotResp = %v, want %v", gotResp, tt.wantResp)
-			}
-			assert.NotEmpty(t, s.results)
-			assert.Equal(t, 2, hookCallCounter)
-		})
-	}
-}
-
-func TestStoreAssessmentResult(t *testing.T) {
-	type args struct {
-		in0        context.Context
-		assessment *orchestrator.StoreAssessmentResultRequest
-	}
-	tests := []struct {
-		name     string
-		args     args
-		wantResp *orchestrator.StoreAssessmentResultResponse
-		wantErr  assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Store assessment to the map",
-			args: args{
-				in0: context.TODO(),
-				assessment: &orchestrator.StoreAssessmentResultRequest{
-					Result: &assessment.AssessmentResult{
-						Id:             testdata.MockAssessmentResultID,
-						MetricId:       "assessmentResultMetricID",
-						EvidenceId:     testdata.MockEvidenceID,
-						CloudServiceId: testdata.MockCloudServiceID,
-						Timestamp:      timestamppb.Now(),
-						MetricConfiguration: &assessment.MetricConfiguration{
-							TargetValue:    toStruct(1.0),
-							Operator:       "<=",
-							IsDefault:      true,
-							CloudServiceId: testdata.MockCloudServiceID,
-							MetricId:       testdata.MockMetricID,
-						},
-						NonComplianceComments: testdata.MockAssessmentResultNonComplianceComment,
-						Compliant:             true,
-						ResourceId:            testdata.MockResourceID,
-						ResourceTypes:         []string{"ResourceType"},
-					},
-				},
-			},
-			wantErr:  assert.NoError,
-			wantResp: &orchestrator.StoreAssessmentResultResponse{},
-		},
-		{
-			name: "Store assessment without metricId to the map",
-			args: args{
-				in0: context.TODO(),
-				assessment: &orchestrator.StoreAssessmentResultRequest{
-					Result: &assessment.AssessmentResult{
-						Id:             testdata.MockAssessmentResultID,
-						EvidenceId:     testdata.MockEvidenceID,
-						CloudServiceId: testdata.MockCloudServiceID,
-						Timestamp:      timestamppb.Now(),
-						MetricConfiguration: &assessment.MetricConfiguration{
-							TargetValue:    toStruct(1.0),
-							Operator:       "<=",
-							IsDefault:      true,
-							CloudServiceId: testdata.MockCloudServiceID,
-							MetricId:       testdata.MockMetricID,
-						},
-						NonComplianceComments: testdata.MockAssessmentResultNonComplianceComment,
-						Compliant:             true,
-						ResourceId:            testdata.MockResourceID,
-						ResourceTypes:         []string{"ResourceType"},
-					},
-				},
-			},
-			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "caused by: invalid AssessmentResult.MetricId: value length must be at least 1 runes")
-			},
-			wantResp: nil,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			s := NewService()
-			gotResp, err := s.StoreAssessmentResult(tt.args.in0, tt.args.assessment)
-			tt.wantErr(t, err)
-			assert.Equal(t, tt.wantResp, gotResp)
-
-			if err == nil {
-				assert.NotNil(t, s.results[testdata.MockAssessmentResultID])
-			} else {
-				assert.Empty(t, s.results)
-			}
-		})
-	}
-}
-
-func TestStoreAssessmentResults(t *testing.T) {
-	const (
-		count1 = 1
-		count2 = 2
-	)
-
-	type fields struct {
-		countElementsInMock    int
-		countElementsInResults int
-	}
-
-	type args struct {
-		streamToServer            *mockStreamer
-		streamToClientWithSendErr *mockStreamerWithSendErr
-		streamToServerWithRecvErr *mockStreamerWithRecvErr
-	}
-
-	tests := []struct {
-		name            string
-		fields          fields
-		args            args
-		wantErr         bool
-		wantRespMessage []orchestrator.StoreAssessmentResultResponse
-		wantErrMessage  string
-	}{
-		{
-			name: "Store 2 assessment results to the map",
-			fields: fields{
-				countElementsInMock:    count2,
-				countElementsInResults: count2,
-			},
-			args:    args{streamToServer: createMockStream(createStoreAssessmentResultRequestsMock(count2))},
-			wantErr: false,
-			wantRespMessage: []orchestrator.StoreAssessmentResultResponse{
-				{
-					Status: true,
-				},
-				{
-					Status: true,
-				},
-			},
-		},
-		{
-			name: "Missing MetricID",
-			fields: fields{
-				countElementsInMock:    count1,
-				countElementsInResults: 0,
-			},
-			args:    args{streamToServer: createMockStream(createStoreAssessmentResultRequestMockWithMissingMetricID(count1))},
-			wantErr: false,
-			wantRespMessage: []orchestrator.StoreAssessmentResultResponse{
-				{
-					Status:        false,
-					StatusMessage: "MetricId: value length must be at least 1 runes",
-				},
-			},
-		},
-		{
-			name: "Error in stream to server - Recv()-err",
-			fields: fields{
-				countElementsInMock:    count1,
-				countElementsInResults: 0,
-			},
-			args:           args{streamToServerWithRecvErr: createMockStreamWithRecvErr(createStoreAssessmentResultRequestsMock(count1))},
-			wantErr:        true,
-			wantErrMessage: "rpc error: code = Unknown desc = cannot receive stream request",
-		},
-		{
-			name: "Error in stream to client - Send()-err",
-			fields: fields{
-				countElementsInMock:    count1,
-				countElementsInResults: 0,
-			},
-			args:           args{streamToClientWithSendErr: createMockStreamWithSendErr(createStoreAssessmentResultRequestsMock(count1))},
-			wantErr:        true,
-			wantErrMessage: "rpc error: code = Unknown desc = cannot stream response to the client",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			s := NewService()
-
-			var err error
-
-			if tt.args.streamToServer != nil {
-				err = s.StoreAssessmentResults(tt.args.streamToServer)
-			} else if tt.args.streamToClientWithSendErr != nil {
-				err = s.StoreAssessmentResults(tt.args.streamToClientWithSendErr)
-			} else if tt.args.streamToServerWithRecvErr != nil {
-				err = s.StoreAssessmentResults(tt.args.streamToServerWithRecvErr)
-			}
-
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Got StoreAssessmentResults() error = %v, wantErrMessage %v", err, tt.wantErr)
-				assert.Equal(t, tt.fields.countElementsInResults, len(s.results))
-				return
-			} else if tt.wantErr {
-				assert.Contains(t, err.Error(), tt.wantErrMessage)
-			} else {
-				// Close stream for testing
-				close(tt.args.streamToServer.SentFromServer)
-				assert.Nil(t, err)
-				assert.Equal(t, tt.fields.countElementsInResults, len(s.results))
-
-				// Check all stream responses from server to client
-				i := 0
-				for elem := range tt.args.streamToServer.SentFromServer {
-					assert.Contains(t, elem.StatusMessage, tt.wantRespMessage[i].StatusMessage)
-					assert.Equal(t, tt.wantRespMessage[i].Status, elem.Status)
-					i++
-				}
-			}
-
-		})
-	}
-}
-
-// createStoreAssessmentResultRequestMockWithMissingMetricID create one StoreAssessmentResultRequest without ToolID
-func createStoreAssessmentResultRequestMockWithMissingMetricID(count int) []*orchestrator.StoreAssessmentResultRequest {
-	req := createStoreAssessmentResultRequestsMock(count)
-
-	req[0].Result.MetricId = ""
-
-	return req
-}
-
-// createStoreAssessmentResultrequestMocks creates store assessment result requests with random assessment result IDs
-func createStoreAssessmentResultRequestsMock(count int) []*orchestrator.StoreAssessmentResultRequest {
-	var mockRequests []*orchestrator.StoreAssessmentResultRequest
-
-	for i := 0; i < count; i++ {
-		storeAssessmentResultRequest := &orchestrator.StoreAssessmentResultRequest{
-			Result: &assessment.AssessmentResult{
-				Id:             uuid.NewString(),
-				MetricId:       fmt.Sprintf("assessmentResultMetricID-%d", i),
-				EvidenceId:     testdata.MockEvidenceID,
-				CloudServiceId: testdata.MockCloudServiceID,
-				Timestamp:      timestamppb.Now(),
-				MetricConfiguration: &assessment.MetricConfiguration{
-					TargetValue:    toStruct(1.0),
-					Operator:       "<=",
-					IsDefault:      true,
-					CloudServiceId: testdata.MockCloudServiceID,
-					MetricId:       testdata.MockMetricID,
-				},
-				NonComplianceComments: testdata.MockAssessmentResultNonComplianceComment,
-				Compliant:             true,
-				ResourceId:            testdata.MockResourceID,
-				ResourceTypes:         []string{"ResourceType"},
-			},
-		}
-
-		mockRequests = append(mockRequests, storeAssessmentResultRequest)
-	}
-
-	return mockRequests
-}
-
-type mockStreamer struct {
-	grpc.ServerStream
-	RecvToServer   chan *orchestrator.StoreAssessmentResultRequest
-	SentFromServer chan *orchestrator.StoreAssessmentResultResponse
-}
-
-func createMockStream(requests []*orchestrator.StoreAssessmentResultRequest) *mockStreamer {
-	m := &mockStreamer{
-		RecvToServer: make(chan *orchestrator.StoreAssessmentResultRequest, len(requests)),
-	}
-	for _, req := range requests {
-		m.RecvToServer <- req
-	}
-
-	m.SentFromServer = make(chan *orchestrator.StoreAssessmentResultResponse, len(requests))
-	return m
-}
-
-func (m mockStreamer) Send(response *orchestrator.StoreAssessmentResultResponse) error {
-	m.SentFromServer <- response
-	return nil
-}
-
-func (m mockStreamer) Recv() (*orchestrator.StoreAssessmentResultRequest, error) {
-	if len(m.RecvToServer) == 0 {
-		return nil, io.EOF
-	}
-	req, more := <-m.RecvToServer
-	if !more {
-		return nil, errors.New("empty")
-	}
-
-	return req, nil
-}
-
-func (mockStreamer) SendAndClose(_ *emptypb.Empty) error {
-	return nil
-}
-
-func (mockStreamer) SetHeader(_ metadata.MD) error {
-	panic("implement me")
-}
-
-func (mockStreamer) SendHeader(_ metadata.MD) error {
-	panic("implement me")
-}
-
-func (mockStreamer) SetTrailer(_ metadata.MD) {
-	panic("implement me")
-}
-
-func (mockStreamer) Context() context.Context {
-	panic("implement me")
-}
-
-func (mockStreamer) SendMsg(_ interface{}) error {
-	panic("implement me")
-}
-
-func (mockStreamer) RecvMsg(_ interface{}) error {
-	panic("implement me")
-}
-
-type mockStreamerWithSendErr struct {
-	grpc.ServerStream
-	RecvToServer   chan *orchestrator.StoreAssessmentResultRequest
-	SentFromServer chan *orchestrator.StoreAssessmentResultResponse
-}
-
-func (mockStreamerWithSendErr) Send(*orchestrator.StoreAssessmentResultResponse) error {
-	return errors.New("Send()-err")
-}
-
-func (m mockStreamerWithSendErr) Recv() (*orchestrator.StoreAssessmentResultRequest, error) {
-	if len(m.RecvToServer) == 0 {
-		return nil, io.EOF
-	}
-	req, more := <-m.RecvToServer
-	if !more {
-		return nil, errors.New("empty")
-	}
-
-	return req, nil
-}
-
-func createMockStreamWithSendErr(requests []*orchestrator.StoreAssessmentResultRequest) *mockStreamerWithSendErr {
-	m := &mockStreamerWithSendErr{
-		RecvToServer: make(chan *orchestrator.StoreAssessmentResultRequest, len(requests)),
-	}
-	for _, req := range requests {
-		m.RecvToServer <- req
-	}
-
-	m.SentFromServer = make(chan *orchestrator.StoreAssessmentResultResponse, len(requests))
-	return m
-}
-
-type mockStreamerWithRecvErr struct {
-	grpc.ServerStream
-	RecvToServer   chan *orchestrator.StoreAssessmentResultRequest
-	SentFromServer chan *orchestrator.StoreAssessmentResultResponse
-}
-
-func (mockStreamerWithRecvErr) Send(*orchestrator.StoreAssessmentResultResponse) error {
-	panic("implement me")
-}
-
-func (mockStreamerWithRecvErr) Recv() (*orchestrator.StoreAssessmentResultRequest, error) {
-	err := errors.New("Recv()-error")
-
-	return nil, err
-}
-
-func createMockStreamWithRecvErr(requests []*orchestrator.StoreAssessmentResultRequest) *mockStreamerWithRecvErr {
-	m := &mockStreamerWithRecvErr{
-		RecvToServer: make(chan *orchestrator.StoreAssessmentResultRequest, len(requests)),
-	}
-	for _, req := range requests {
-		m.RecvToServer <- req
-	}
-
-	m.SentFromServer = make(chan *orchestrator.StoreAssessmentResultResponse, len(requests))
-	return m
 }
 
 func toStruct(f float32) (s *structpb.Value) {
